+++ HttpPool.cs
@@ -4,6 +4,8 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 
+namespace ShopServices;
+
 internal static class HttpPool
 {
 	public const byte Limit = 16;