+++ Search/Searchers/K4PartsSearcher.cs
@@ -5,10 +5,11 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json.Linq;
+using ShopServices.Search.Models;
 
-namespace ShopServices.Functions.Search;
+namespace ShopServices.Search.Searchers;
 
-public class K4PartsSearcher : Searcher
+internal class K4PartsSearcher : Searcher
 {
 	private const int Limit = 8;